cleanup: Splits the IndexService class up into PFIBuilder.php
Splits the class into two to make the code cleaner. This is part of the cleanup process.
